requirements.txt
numpy==1.26.4
numba==0.60.0
joblib==1.4.2
imbalanced-learn==0.12.4
dcor==0.6
gplearn==0.4.2
matplotlib==3.9.4
optuna==4.2.1
pandas==2.2.3
pyHSICLasso==1.4.2
scipy==1.13.1
scikit-learn==1.5.2
seaborn==0.13.2
sklearn-pandas==2.2.0
umap-learn==0.5.7
xgboost==2.1.4







